FIX: Prevent race condition in recurring automations (#26828)
Recurring automations are triggered by a scheduled job that runs every minute and checks for due automations, runs them and then marks as them as completed (by deleting the `PendingAutomation` record). However, the job is currently subject to a race condition where a recurring automation can be executed more than once at its due date if it takes more than a minute to finish.

This commit adds a mutex around the code that triggers the recurring automation so that no concurrent executions can happen for a single automation.

Meta topic: https://meta.discourse.org/t/daily-summary-9pm-utc/291850/119?u=osama.

+++ plugins/automation/app/jobs/scheduled/discourse_automation_tracker.rb
@@ -23,22 +23,36 @@ module Jobs
     end
 
     def send_pending_pm(pending_pm)
-      DiscourseAutomation::Scriptable::Utils.send_pm(
-        pending_pm.attributes.slice("target_usernames", "title", "raw"),
-        sender: pending_pm.sender,
-        prefers_encrypt: pending_pm.prefers_encrypt,
-      )
+      DistributedMutex.synchronize(
+        "automation_send_pending_pm_#{pending_pm.id}",
+        validity: 30.minutes,
+      ) do
+        next if !DiscourseAutomation::PendingPm.exists?(pending_pm.id)
 
-      pending_pm.destroy!
+        DiscourseAutomation::Scriptable::Utils.send_pm(
+          pending_pm.attributes.slice("target_usernames", "title", "raw"),
+          sender: pending_pm.sender,
+          prefers_encrypt: pending_pm.prefers_encrypt,
+        )
+
+        pending_pm.destroy!
+      end
     end
 
     def run_pending_automation(pending_automation)
-      pending_automation.automation.trigger!(
-        "kind" => pending_automation.automation.trigger,
-        "execute_at" => pending_automation.execute_at,
-      )
+      DistributedMutex.synchronize(
+        "process_pending_automation_#{pending_automation.id}",
+        validity: 30.minutes,
+      ) do
+        next if !DiscourseAutomation::PendingAutomation.exists?(pending_automation.id)
 
-      pending_automation.destroy!
+        pending_automation.automation.trigger!(
+          "kind" => pending_automation.automation.trigger,
+          "execute_at" => pending_automation.execute_at,
+        )
+
+        pending_automation.destroy!
+      end
     end
   end
 end
